Closes feature/can-rx-perf (#11)
This commit decouples the kvaser message receive callback from the message handling in the llserver. This is done by adding a new thread and a queue. Messagesa are added to the queue and removed/processed as soon as possible by a new message handling thread. Previously the long message parsing lead to sampling rate dropoffs. The new version appears to have fixed the bug

Additionally this commit fixes a memory leak in Node.cpp