Bugfix/llserver freeze up (#13)
Fixes a bug, where the whole llserver would freeze up for ~10 seconds. Ultimately the influxdb is responsible for this, since it sometimes takes >200ms to respond to an insert, instead of the normal 2-20. This PR fixes it by **completely** decoupling the sending thread from the rest of the llserver. I believe that this is also the root cause behind #11.